requirements.txt
fastapi
uvicorn[standard]
httpx
torch
transformers
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi 
azure.monitor.opentelemetry
python-dotenv
accelerate